+++ config/configmapprovider/provider.go
@@ -12,15 +12,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package config // import "go.opentelemetry.io/collector/config"
+package configmapprovider // import "go.opentelemetry.io/collector/config/configmapprovider"
 
 import (
 	"context"
+
+	"go.opentelemetry.io/collector/config"
 )
 
-// MapProvider is an interface that helps providing configuration's parser.
+// Provider is an interface that helps providing configuration's parser.
 // Implementations may load the parser from a file, a database or any other source.
-type MapProvider interface {
+type Provider interface {
 	// Retrieve goes to the configuration source and retrieves the selected data which
 	// contains the value to be injected in the configuration and the corresponding watcher that
 	// will be used to monitor for updates of the retrieved value.
@@ -35,7 +37,7 @@ type MapProvider interface {
 // Retrieved holds the result of a call to the Retrieve method of a Session object.
 type Retrieved interface {
 	// Get returns the Map.
-	Get() *Map
+	Get() *config.Map
 }
 
 // WatchableRetrieved is an extension for Retrieved that if implemented,